+++ src/BasicTypes/HashMaps/HashMap.spec.ts
@@ -1433,4 +1433,161 @@ describe("HashMap()", () => {
             expect(input).eql(unit);
         });
     });
+
+    describe(".flatMap()", () => {
+        it("applies the callback to all properties of the input object", () => {
+            // ----------------------------------------------------------------
+            // explain your test
+
+            // this test proves that the callback function (the mapper) is
+            // run against all properties of the input object
+
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit = {
+                a: { z: [ 1, 2, 3 ] },
+                b: { y: [ 4, 5, 6 ] },
+            }
+            const expectedResult = {
+                z: [ 2, 4, 6 ],
+                y: [ 8, 10, 12 ],
+            }
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            const actualResult = HashMap.flatMap(
+                unit,
+                (nestedObj) =>
+                    HashMap.map(
+                        nestedObj,
+                        (value) => value.map((num) => num * 2)
+                    )
+
+            );
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).to.eql(expectedResult);
+        });
+
+        it("flattens the result of the map operation by one level", () => {
+            // ----------------------------------------------------------------
+            // explain your test
+
+            // this test proves that the property keys in the returned object
+            // are the keys from the nested objects
+
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit = {
+                a: { z: [ 1, 2, 3 ] },
+                b: { y: [ 4, 5, 6 ] },
+            }
+            const expectedResult = {
+                z: [ 2, 4, 6 ],
+                y: [ 8, 10, 12 ],
+            }
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            const actualResult = HashMap.flatMap(
+                unit,
+                (nestedObj) =>
+                    HashMap.map(
+                        nestedObj,
+                        (value) => value.map((num) => num * 2)
+                    )
+
+            );
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).to.eql(expectedResult);
+        });
+
+        it("does not modify the input object", () => {
+            // ----------------------------------------------------------------
+            // explain your test
+
+            // this test proves that the input object is left untouched
+            // by the flatMap() method
+
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit = {
+                a: { z: [ 1, 2, 3 ] },
+                b: { y: [ 4, 5, 6 ] },
+            }
+
+            // create a copy of out input object, to compare against
+            // afterward
+            const expectedResult = {
+                ...unit
+            }
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            const actualResult = HashMap.flatMap(
+                unit,
+                (nestedObj) =>
+                    HashMap.map(
+                        nestedObj,
+                        (value) => value.map((num) => num * 2)
+                    )
+
+            );
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(actualResult).to.not.eql(expectedResult);
+            expect(unit).to.eql(expectedResult);
+        });
+
+        it("only supports nested objects", () => {
+            // ----------------------------------------------------------------
+            // explain your test
+
+            // this test proves that HashMap.flatMap() will throw an error
+            // if we pass in an array
+
+            // ----------------------------------------------------------------
+            // setup your test
+
+            const unit = {
+                a: [ 1, 2, 3 ],
+                b: [ 4, 5, 6 ],
+            }
+
+
+            // ----------------------------------------------------------------
+            // perform the change
+
+            let caughtError = false;
+            let actualResult;
+            try {
+                actualResult = HashMap.flatMap(
+                    unit,
+                    (nestedObj) => nestedObj
+                );
+            }
+            catch (e) {
+                caughtError = true;
+            }
+
+            // ----------------------------------------------------------------
+            // test the results
+
+            expect(caughtError).to.be.true;
+        });
+
+    });
 });